Move the add court document functionality to the top
